third_party/blink/web_tests/platform/generic/external/wpt/css/css-page/page-rule-declarations-004-expected.txt

This is a testharness.js-based test.
FAIL rule with invalid selector ['--a'] assert_not_equals: no @page rule should have been parsed got disallowed value 6
PASS adding a blank @page rule
PASS assigning invalid selector text ['1']
PASS assigning invalid selector text ['-3']
FAIL assigning invalid selector text ['--a'] assert_equals: should not be able to assign an invalid selector expected "" but got "--a"
PASS assigning invalid selector text ['7cm']
PASS assigning invalid selector text ['0.17']
PASS assigning invalid selector text ['a, 123']
Harness: the test ran to completion.
